@@ -2,32 +2,32 @@
   "name": "robonaut-server-integration-test",
   "private": true,
   "dependencies": {
-    "axios": "0.25.0",
-    "pg": "8.7.1"
+    "axios": "0.27.2",
+    "pg": "8.7.3 "
   },
   "devDependencies": {
-    "@tsconfig/node16": "1.0.2",
-    "@types/jest": "27.4.0",
+    "@tsconfig/node16": "1.0.3",
+    "@types/jest": "28.1.2",
     "@types/mkdirp": "1.0.2",
-    "@types/node": "16.11.22",
+    "@types/node": "ts3.7",
     "@types/rimraf": "3.0.2",
-    "@types/pg": "8.6.4",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
-    "eslint": "8.8.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-jest": "26.0.0",
+    "@types/pg": "8.6.5",
+    "@typescript-eslint/eslint-plugin": "5.28.0",
+    "@typescript-eslint/parser": "5.28.0",
+    "eslint": "8.17.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-plugin-jest": "26.5.3",
     "eslint-plugin-prettier": "4.0.0",
-    "jest": "27.4.7",
-    "jest-environment-node": "27.4.6",
-    "jest-junit": "13.0.0",
+    "jest": "28.1.1",
+    "jest-environment-node": "28.1.1",
+    "jest-junit": "13.2.0",
     "mkdirp": "1.0.4",
-    "prettier": "2.5.1",
+    "prettier": "2.7.1",
     "rimraf": "3.0.2",
-    "testcontainers": "8.2.0",
-    "ts-jest": "27.1.3",
-    "ts-node": "10.4.0",
-    "typescript": "4.5.5"
+    "testcontainers": "8.10.1",
+    "ts-jest": "28.0.5",
+    "ts-node": "10.8.1",
+    "typescript": "4.7.3"
   },
   "scripts": {
     "lint": "eslint --ignore-path .gitignore .",
